introduce interface classes
