+++ audbackend/__init__.py
@@ -4,9 +4,10 @@ from audbackend.core.api import create
 from audbackend.core.api import delete
 from audbackend.core.api import register
 from audbackend.core.backend import Backend
-from audbackend.core.backend import VersionedBackend
 from audbackend.core.errors import BackendError
 from audbackend.core.filesystem import FileSystem
+from audbackend.core.interface import Unversioned
+from audbackend.core.interface import Versioned
 from audbackend.core.repository import Repository
 
 # Import optional backends
